requirements.txt
google-auth==2.29.0
requests==2.31.0